+++ packages/client.api/src/Definitions.ts
@@ -15,7 +15,7 @@
  */
 
 import type { ObjectTypePropertyDefinition } from "@osdk/api";
-import type { PropertyValueWireToClient } from "@osdk/client.api";
+import type { PropertyValueWireToClient } from "./mapping/PropertyValueMapping.js";
 
 type MaybeArray<T extends { multiplicity?: boolean | undefined }, U> =
   T["multiplicity"] extends true ? Array<U> : U;
