+++ packages/client.api/src/aggregate/AggregationsResults.ts
@@ -15,7 +15,8 @@
  */
 
 import type { ObjectOrInterfaceDefinition } from "@osdk/api";
-import type { AggregatableKeys, AggregateOpts } from "@osdk/client.api";
+import type { AggregatableKeys } from "./AggregatableKeys.js";
+import type { AggregateOpts } from "./AggregateOpts.js";
 import type { AggregationResultsWithGroups } from "./AggregationResultsWithGroups.js";
 import type {
   AggregationCountResult,
