+++ packages/client/src/internal/conversions/legacyToModernSingleAggregationResult.ts
@@ -14,10 +14,12 @@
  * limitations under the License.
  */
 
-import type { AggregationClause } from "@osdk/client.api";
+import type {
+  AggregationClause,
+  AggregationResultsWithoutGroups,
+} from "@osdk/client.api";
 import type { AggregateObjectsResponseV2 } from "@osdk/internal.foundry";
 import invariant from "tiny-invariant";
-import type { AggregationResultsWithoutGroups } from "../../query/aggregations/AggregationResultsWithoutGroups.js";
 import type { ArrayElement } from "../../util/ArrayElement.js";
 
 export function legacyToModernSingleAggregationResult<
