+++ packages/client/src/object/aggregate.ts
@@ -15,7 +15,12 @@
  */
 
 import type { ObjectOrInterfaceDefinition } from "@osdk/api";
-import type { AggregateOpts, AggregateOptsThatErrors } from "@osdk/client.api";
+import type {
+  AggregateOpts,
+  AggregateOptsThatErrors,
+  AggregationResultsWithGroups,
+  AggregationsResults,
+} from "@osdk/client.api";
 import type {
   AggregateObjectsRequestV2,
   AggregateObjectsResponseV2,
@@ -28,8 +33,6 @@ import { modernToLegacyAggregationClause } from "../internal/conversions/modernT
 import { modernToLegacyGroupByClause } from "../internal/conversions/modernToLegacyGroupByClause.js";
 import { modernToLegacyWhereClause } from "../internal/conversions/modernToLegacyWhereClause.js";
 import type { MinimalClient } from "../MinimalClientContext.js";
-import type { AggregationResultsWithGroups } from "../query/aggregations/AggregationResultsWithGroups.js";
-import type { AggregationsResults } from "../query/aggregations/AggregationsResults.js";
 import { addUserAgent } from "../util/addUserAgent.js";
 import type { ArrayElement } from "../util/ArrayElement.js";
